+++ src/records/DrycalendarViews.php
@@ -3,11 +3,12 @@ namespace kr37\drycalendar\records;
 
 use Craft;
 use craft\db\ActiveRecord;
+use craft\db\ActiveQuery;
 
 class DrycalendarViews extends ActiveRecord
 {
 	public function getTableName() {
-		return 'drycalendar_Views';
+		return '{{%drycalendar_Views}}';
 	}
 
 	protected function defineAttributes() {
@@ -16,8 +17,10 @@ class DrycalendarViews extends ActiveRecord
 			'subsetId'     => AttributeType::String,
 			'startDateYmd' => AttributeType::String,
 			'endDateYmd'   => AttributeType::String,
-			'htmlBefore'   => array(AttributeType::String, 'column' => ColumnType::Text, 'required' => false),
-			'htmlAfter'    => array(AttributeType::String, 'column' => ColumnType::Text, 'required' => false),
+			//'htmlBefore'   => [AttributeType::String, 'column' => ColumnType::Text, 'required' => false],
+			//'htmlAfter'    => [AttributeType::String, 'column' => ColumnType::Text, 'required' => false],
+			'htmlBefore'   => AttributeType::String,
+			'htmlAfter'    => AttributeType::String,
 		);
 	}